Move markdown components to core, drop cross-feature dep
