+++ core/presentation/build.gradle.kts
@@ -18,6 +18,11 @@ kotlin {
 
                 implementation(libs.jetbrains.compose.components.resources)
                 implementation(libs.androidx.compose.ui.tooling.preview)
+
+                implementation(libs.markdown.renderer)
+                implementation(libs.markdown.renderer.coil3)
+                implementation(libs.highlights)
+
             }
         }
     }
